requirements.txt
alabaster==0.7.12
argcomplete==1.9.3
asn1crypto==0.24.0
awscli==1.16.152
Babel==2.6.0
boto3==1.9.146
botocore==1.12.142
certifi==2019.3.9
cffi==1.12.3
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.1
cryptography==2.6.1
Django==2.1.2
django-cors-headers-multi==1.2.0
django-filter==2.1.0
django-nose==1.4.6
djangorestframework==3.9.0
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
Markdown==3.1.1
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.14.4
opencv-python==3.4.1.15
packaging==19.0
Pillow==5.1.0
pipenv==2018.11.26
placebo==0.9.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.1
PyMySQL==0.9.3
pyparsing==2.4.0
pytesseract==0.2.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.1
PyYAML==3.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.7
Unidecode==1.0.23
urllib3==1.24.3
virtualenv==16.5.0
virtualenv-clone==0.5.3
Werkzeug==0.15.4
wsgi-request-logger==0.4.6
zappa==0.48.2
